# Add OpenMP-accelerated sparse matrix operations for Hessian computation